2024年8月9日 14点33分

+++ SocketTransferServer/src/main/java/com/mdsd/cloud/response/ResponseTy.java
@@ -17,7 +17,7 @@ public class ResponseTy<T> implements Serializable {
     /**
      * 0为成功,其它为失败
      */
-    private short state;
+    private int state;
 
     /**
      * 操作描述,当state不为0时参见此值
@@ -27,7 +27,7 @@ public class ResponseTy<T> implements Serializable {
     /**
      * 记录总数,在列表接口返回时有此值
      */
-    private Integer totalNum;
+    private int totalNum;
 
     /**
      * 实际返回数,可能为 list 或 map
